refactor: move code into __init__ of package
